kas: migrate from deprecated refspec syntax
Causes the following error messages:

Using deprecated refspec for repository "meta-webkit".
You should migrate to commit/tag/branch.

Note:

If we use only "tag", we get the following warning:

Using tag without commit for repository "poky" is unsafe as tags are mutable.

According to kas reference documentation, it is advised to put both
"tag" and "commit" when using tags:

If a commit was specified, kas checks that the tag points to this
commit.

Signed-off-by: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>