bsp: rockchip: linux: cleanup quartz64-a ETH config
This is the minimal config I was able to get that enables wire
networking for Quartz64-a. The initial config was got from one of
pipelines here https://gitlab.com/pgwipeout/quartz64_ci/-/pipelines .
Note that this might not be the absolute minimal config needed, so any
suggestions are welcome.

Signed-off-by: Daniil Klimuk <daniil.klimuk@3mdeb.com>